+++ cli/src/commands/parse.rs
@@ -7,7 +7,6 @@ use std::{fs, io};
 use anyhow::{anyhow, Context, Result};
 use phylum_lockfile::{LockfileFormat, Package, PackageVersion, Parse, ThirdPartyVersion};
 use phylum_types::types::package::{PackageDescriptor, PackageDescriptorAndLockfile};
-use walkdir::WalkDir;
 
 use crate::commands::{CommandResult, ExitCode};
 use crate::{config, print_user_warning};
@@ -181,11 +180,10 @@ fn find_manifest_format(path: &Path) -> Option<(LockfileFormat, Option<PathBuf>)
 
     // Look for formats which already have a lockfile generated.
     let manifest_lockfile = formats.find_map(|format| {
-        let manifest_lockfile = WalkDir::new(manifest_dir)
-            .into_iter()
-            .flatten()
-            .find(|entry| format.parser().is_path_lockfile(entry.path()))?;
-        Some((format, manifest_lockfile.path().to_owned()))
+        let manifest_lockfile = find_direntry_upwards::<32, _>(manifest_dir, |path| {
+            format.parser().is_path_lockfile(path)
+        })?;
+        Some((format, manifest_lockfile))
     });
 
     // Return existing lockfile or format capable of generating it.
@@ -251,8 +249,36 @@ fn filter_packages(mut packages: Vec<Package>) -> Vec<PackageDescriptor> {
         .collect()
 }
 
+/// Find a file by walking from a directory towards the root.
+///
+/// `MAX_DEPTH` is the maximum number of directory traversals before the search
+/// will be abandoned. A `MAX_DEPTH` of `0` will only search the `origin`
+/// directory.
+fn find_direntry_upwards<const MAX_DEPTH: usize, P>(
+    mut origin: &Path,
+    mut predicate: P,
+) -> Option<PathBuf>
+where
+    P: FnMut(&Path) -> bool,
+{
+    for _ in 0..=MAX_DEPTH {
+        for entry in fs::read_dir(origin).ok()?.flatten() {
+            let path = entry.path();
+            if predicate(&path) {
+                return Some(path);
+            }
+        }
+
+        origin = origin.parent()?;
+    }
+
+    None
+}
+
 #[cfg(test)]
 mod tests {
+    use std::fs::{self, File};
+
     use super::*;
 
     #[test]
@@ -287,5 +313,48 @@ mod tests {
             let parsed = try_get_packages(PathBuf::from(file)).unwrap();
             assert_eq!(parsed.format, expected_format, "{}", file);
         }
+    }
+
+    #[test]
+    fn find_lockfile_for_manifest() {
+        let tempdir = tempfile::tempdir().unwrap();
+        let tempdir = tempdir.path().canonicalize().unwrap();
+
+        // Create manifest.
+        let manifest_dir = tempdir.join("manifest");
+        let manifest_path = manifest_dir.join("Cargo.toml");
+        fs::create_dir_all(&manifest_dir).unwrap();
+        File::create(&manifest_path).unwrap();
+
+        // Ensure lockfiles below manifest are ignored.
+
+        // Create lockfile below manifest.
+        let child_lockfile_dir = manifest_dir.join("sub");
+        fs::create_dir_all(&child_lockfile_dir).unwrap();
+        File::create(child_lockfile_dir.join("Cargo.lock")).unwrap();
+
+        let (_, path) = find_manifest_format(&manifest_path).unwrap();
+
+        assert_eq!(path, None);
+
+        // Accept lockfiles above the manifest.
+
+        // Create lockfile above manifest.
+        let parent_lockfile_path = tempdir.join("Cargo.lock");
+        File::create(&parent_lockfile_path).unwrap();
+
+        let (_, path) = find_manifest_format(&manifest_path).unwrap();
+
+        assert_eq!(path, Some(parent_lockfile_path));
+
+        // Prefer lockfiles "closer" to the manifest.
+
+        // Create lockfile in the manifest directory.
+        let sibling_lockfile_path = manifest_dir.join("Cargo.lock");
+        File::create(&sibling_lockfile_path).unwrap();
+
+        let (_, path) = find_manifest_format(&manifest_path).unwrap();
+
+        assert_eq!(path, Some(sibling_lockfile_path));
     }
 }