+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Include lockfile paths when analyzing projects
 
+### Fixed
+- Detect lockfile in current path before searching in subdirectories
+
 ## [5.5.0] - 2023-07-18
 
 ### Added