+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- Generating lockfiles for `*.csproj` files
+
 ### Changed
 - Include lockfile paths when analyzing projects
 - Generate and use API Keys instead of OpenID Connect tokens